+++ mathinspector/__main__.py
@@ -1,4 +1,4 @@
-from . import main
+from .app import main
 
 
 if __name__ == '__main__':
