+++ setup.cfg
@@ -68,6 +68,6 @@ packages=find:
 python_requires=>=3.6
 [options.entry_points]
 console_scripts=
-    mathinspector=mathinspector:main
+    mathinspector=mathinspector.app:main
 [options.package_data]
 mathinspector=assets/*, assets/*/*
